V1.0 official release
I was working on a branch without comitting to git for a while, and finished the game in that time.
Today i picked the project up again so i could work on some final changes and tweaks to deliver the game, as the deadline is around the corner.
Everything should be working as intended, we're out of Beta, this is the official release of the game